Parse as a float instead of an int

+++ client/packages/system/src/Stock/Components/InventoryAdjustment/InventoryAdjustmentModal.tsx
@@ -144,7 +144,7 @@ export const InventoryAdjustmentModal: FC<InventoryAdjustmentModalProps> = ({
               <NumericTextInput
                 width={INPUT_WIDTH}
                 disabled={true}
-                value={parseInt(
+                value={parseFloat(
                   (
                     stockLine.totalNumberOfPacks +
                     (draft.adjustmentType === AdjustmentTypeInput.Addition
